# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Tienda Online API",
        "description": "API REST para gestión de tienda Vue + Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1": {},
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "6796c44e1f6079c411abbfed066c5907",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan López"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "juan@tienda.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "operationId": "da2ca4d7a5cdec77b8b45a20373535cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@tienda.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso"
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener usuario autenticado",
                "operationId": "ae2f8c1ae1f9c5aee99c01f5d13da30b",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar productos con filtros y paginación",
                "operationId": "2af0fdbcd7c972b8f9bd45123cd57cfb",
                "parameters": [
                    {
                        "name": "busqueda",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "precio_min",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "precio_max",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 6
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos paginada"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear producto",
                "operationId": "434c8599d4fe9aaa5aee59a5be0cf3fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "precio",
                                    "stock",
                                    "categoria_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Camiseta"
                                    },
                                    "precio": {
                                        "type": "number",
                                        "example": 150
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "imagen": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto",
                "operationId": "7f8c025cedd4418ed0240a21f7f9402a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar producto",
                "operationId": "82ed71c67980b16764a81d295405c031",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "precio": {
                                        "type": "number"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar producto",
                "operationId": "075714328668dc259b9f6efb188ab787",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado"
                    },
                    "403": {
                        "description": "Sin permisos"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}